+++ CHANGELOG.md
@@ -29,7 +29,9 @@
 - The language server now allows extracting the start of a pipeline into a
   variable.
   ([Giacomo Cavalieri](https://github.com/giacomocavalieri))
-- The language server now suggests adding missing generic parameters to custom types.
+
+- The language server now suggests adding missing type parameters
+  to custom generic types.
   ([Andi Pabst](https://github.com/andipabst))
 
 ### Formatter